+++ app/entity/model/Episode.go
@@ -1,9 +1,9 @@
 package model
 
 type Episode struct {
-	EpisodeId uint   `gorm:"primaryKey" json:"episode_id"`
+	EpisodeId int64  `gorm:"primaryKey" json:"episode_id"`
 	Name      string `gorm:"not null" json:"name"`
-	MovieId   uint   `gorm:"index:,not null" json:"movie_id"`
+	MovieId   int64  `gorm:"index:,not null" json:"movie_id"`
 	Status    int    `gorm:"index:,default:1" json:"status"`
 	BaseModel
 }
