+++ app/entity/model/Role.go
@@ -1,7 +1,7 @@
 package model
 
 type Role struct {
-	RoleId uint   `gorm:"primaryKey" json:"role_id"`
+	RoleId int64  `gorm:"primaryKey" json:"role_id"`
 	Name   string `gorm:"not null" json:"name"`
 	Slug   string `gorm:"index:,not null" json:"slug"`
 	Status int    `gorm:"index:,default:1" json:"status"`
