+++ app/entity/model/Tag.go
@@ -1,7 +1,7 @@
 package model
 
 type Tag struct {
-	TagId  uint   `gorm:"primaryKey" json:"tag_id"`
+	TagId  int64  `gorm:"primaryKey" json:"tag_id"`
 	Name   string `gorm:"not null" json:"name"`
 	Slug   string `gorm:"index:,not null" json:"slug"`
 	Status int    `gorm:"index:,default:1" json:"status"`
