+++ app/entity/response/UserResponse.go
@@ -1,7 +1,7 @@
 package response
 
 type UserResponse struct {
-	UserId   uint   `json:"user_id"`
+	UserId   int64  `json:"user_id"`
 	Name     string `json:"name"`
 	Username string `json:"username"`
 	Gender   int    `json:"gender"`
